Updates: - Updated some of the UI for signing in. - Can delete posts from firestore now - Checks current distance from user. Only checks once the posts have been captured though. - Added distance to the map restaurant list sheet. - Started working on the share function. Check out open graph tags - Address button now works. - Gets user avatar now. - Some UI updates. Scrim and fonts. - Profile screen now loads posts only for that user. - Adding a login screen. Currently FB sign in works. - Added Google to login screen. Gotta make it look pretty. - Moving most functions out of ApplicationViewModel.kt, and into their own individual view models. This makes things much smoother and more organized, I feel.
To-Do:
- Make proper sign in screen.
- Fix bottom sheet. (Needs more functions and it should look better. Also should blur or shade the background)
- Make camera look more functional.

Issues:
- Profile screen loads twice. Sing in screen takes over the whole screen, no way to exit and it doesn't leave.
- Images don't load. I think the reason is because the images are loaded, and then it changes to an empty Photo(). Maybe I can solve this by getting rid of the empty photo?
- Permissions is still pretty weird.
- App times out when trying to change screen from maps
- Profile bottom sheet doesn't look right. Need to add items and make it overlay the background.

+++ app/src/main/java/com/example/kbbqreview/screens/HomeScreen/HomeScreenViewModel.kt
@@ -38,9 +38,9 @@ class HomeScreenViewModel : ViewModel() {
                                     .orEmpty(),
                                 location = documentSnapshot.getGeoPoint("location"),
                                 valueMeat = documentSnapshot.getLong("value_meat")!!.toInt(),
-                                valueSideDishes = documentSnapshot.getLong("value_side_dishes")!!,
-                                valueAtmosphere = documentSnapshot.getLong("value_atmosphere")!!,
-                                valueAmenities = documentSnapshot.getLong("value_amenities")!!,
+                                valueSideDishes = documentSnapshot.getLong("value_side_dishes")!!.toInt(),
+                                valueAtmosphere = documentSnapshot.getLong("value_atmosphere")!!.toInt(),
+                                valueAmenities = documentSnapshot.getLong("value_amenities")!!.toInt(),
                                 photoList = getPhotos(firebaseId),
                                 distance = 0.0
                             )
